Separate out publish to a module
Separate out publish to IHaskell.Publish
Replace publish with publishResult, which also requires more arguments.
